introduces scrollback functionality
