+++ src/components/Projects.tsx
@@ -15,6 +15,7 @@ import React, { useCallback } from "react";
 import BuyButton from "./BuyButton";
 import Flex from "./Flex";
 import ProjectsGrid from "./ProjectsGrid";
+import Scrollback from "./Scrollback";
 import styled from "styled-components";
 import theme from "./theme";
 
@@ -23,26 +24,7 @@ const BuyButtonContainer = styled.div<PositionProps & SpaceProps>`
   ${space};
 `;
 
-type ScrollbackProps = PositionProps &
-  TypographyProps &
-  BackgroundProps &
-  LayoutProps;
-
-const Scrollback = styled.button<ScrollbackProps>`
-  ${position};
-  ${typography};
-  ${background};
-  ${layout};
-  border: none;
-`;
-
 const Projects = () => {
-  const scrollToTop = useCallback(() => {
-    if (window) {
-      window.scrollTo({ top: 0, behavior: "smooth" });
-    }
-  }, []);
-
   return (
     <Flex flex="auto" justifyContent="center" alignItems="center">
       <ProjectsGrid />
@@ -54,18 +36,7 @@ const Projects = () => {
         <BuyButton />
       </BuyButtonContainer>
 
-      <Scrollback
-        display={["block", "block", "block", "none"]}
-        position="fixed"
-        bottom={6}
-        left={6}
-        fontSize={[3, 4]}
-        background="transparent"
-        onClick={scrollToTop}
-        zIndex={theme.zIndexes.inFront}
-      >
-        UP
-      </Scrollback>
+      <Scrollback bottom={20} left={20} />
     </Flex>
   );
 };
